fix: restructure directories
